Made routes more modular so it is easier to debug and use
