+++ app.js
@@ -10,6 +10,30 @@ var session = require('client-sessions');
 var index = require('./routes/index');
 var users = require('./routes/users');
 
+//admin routes
+var register_employee_post=require("./routes/admin/register-employee-post");
+var admin_get=require("./routes/admin/admin-get");
+var display_employees_get=require("./routes/admin/display-employees-get");
+
+//project routes
+var display_all_projects=require("./routes/projects/display-all-projects-get");
+var milestone=require("./routes/projects/milestone-get");
+var project_edit=require("./routes/projects/project-edit");
+var project_routing=require("./routes/projects/projects-routing");
+var project_creation=require("./routes/projects/project-creation-post");
+
+//employee routes
+var employee_calendar=require("./routes/employee/employee-calendar");
+var employee_routing=require("./routes/employee/employee-routing");
+
+//all-page-components-routes
+var login=require("./routes/all-page-components/login");
+var logout=require("./routes/all-page-components/logout");
+var notifications=require("./routes/all-page-components/notifications");
+var tasks=require("./routes/all-page-components/tasks");
+var user_info=require("./routes/all-page-components/user-info");
+
+
 var app = express();
 
 app.set('views', path.join(__dirname, 'views'));
@@ -29,8 +53,33 @@ app.use(session({
     activeDuration: 5 * 60 * 1000
 }));
 
-app.use('/', index);
+//routes
+/**
+ * Date Created:31/08/24
+ * Bug: might need to set routes in priority to avoid conflicts
+ */
+app.use('/', login);
+app.use(logout);
+app.use(tasks);
+app.use(notifications);
+app.use(user_info);
+
+app.use( register_employee_post);
+app.use( admin_get);
+app.use( display_employees_get);
+
+app.use( project_creation);
+app.use( display_all_projects);
+app.use( milestone);
+app.use( project_edit);
+app.use( project_routing);
+
+app.use(employee_calendar);
+app.use(employee_routing);
+
 app.use('/users', users);
+app.use(index);
+
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
