+++ past_projects.txt
@@ -1,24 +1,24 @@
-Start Date : Sun Jan 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sun Jan 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Mon Jan 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sat Jan 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Wed Feb 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Wed Feb 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Thu Feb 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Tue Feb 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Wed Mar 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Wed Mar 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Thu Mar 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Tue Mar 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sat Apr 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sat Apr 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sun Apr 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Fri Apr 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Mon May 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Mon May 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Tue May 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sun May 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Thu Jun 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Thu Jun 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Fri Jun 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Wed Jun 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sat Jul 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sat Jul 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sun Jul 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Fri Jul 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Tue Aug 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Tue Aug 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Wed Aug 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Mon Aug 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Fri Sep 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Fri Sep 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sat Sep 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Thu Sep 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sun Oct 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sun Oct 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Mon Oct 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Sat Oct 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Wed Nov 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Wed Nov 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Thu Nov 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Tue Nov 28 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Fri Dec 01 2017 09:50:40 GMT+0200 (SAST) :: End Date : Fri Dec 15 2017 09:50:40 GMT+0200 (SAST)
-Start Date : Sat Dec 16 2017 09:50:40 GMT+0200 (SAST) :: End Date : Thu Dec 28 2017 09:50:40 GMT+0200 (SAST)
+Start Date : Sun Jan 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sun Jan 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Mon Jan 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sat Jan 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Wed Feb 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Wed Feb 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Thu Feb 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Tue Feb 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Wed Mar 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Wed Mar 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Thu Mar 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Tue Mar 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sat Apr 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sat Apr 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sun Apr 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Fri Apr 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Mon May 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Mon May 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Tue May 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sun May 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Thu Jun 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Thu Jun 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Fri Jun 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Wed Jun 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sat Jul 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sat Jul 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sun Jul 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Fri Jul 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Tue Aug 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Tue Aug 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Wed Aug 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Mon Aug 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Fri Sep 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Fri Sep 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sat Sep 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Thu Sep 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sun Oct 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sun Oct 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Mon Oct 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Sat Oct 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Wed Nov 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Wed Nov 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Thu Nov 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Tue Nov 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Fri Dec 01 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Fri Dec 15 2017 18:12:49 GMT+0200 (South Africa Standard Time)
+Start Date : Sat Dec 16 2017 18:12:49 GMT+0200 (South Africa Standard Time) :: End Date : Thu Dec 28 2017 18:12:49 GMT+0200 (South Africa Standard Time)
